% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWATHtoMRM.R
\name{SWATHtoMRM}
\alias{SWATHtoMRM}
\title{SWATHtoMRM}
\usage{
SWATHtoMRM(d.in = ".", d.out = { }, polarity = c("positive", "negative"),
  adduct.list = c("HILIC", "RPLC"), peakwidth = c(8, 30), sn = 20,
  minfrac = 1, minfrac.vote = 0.5, nSlaves = 6, ppm.pd = 15,
  cutoff = 0.8, ppm.ms2.mtp = 15, int.filter.ms1.field = c("maxo", "into",
  "intb"), int.filter.ms1 = NULL, int.filter.ms2 = 200, isFWHM = FALSE,
  is.plot.eic.feature = FALSE, is.plot.eic.spec = FALSE,
  is.ms1.only = FALSE, rerun = FALSE)
}
\arguments{
\item{d.in}{Directory path to load the orgnized data files in mzXML format}

\item{d.out}{Directory path to save analysis results, if not set, the results
will be saved in 'Results(.x)' folder}

\item{polarity}{Polarity setup when acquiring SWATH-MS data}

\item{adduct.list}{Adduct list for CAMERA annotation, which is inherent in
the package}

\item{peakwidth}{Peak width range in seconds when detecting peaks with xcms}

\item{sn}{Signal to noise ratio threshold when detecting peaks with xcms}

\item{minfrac}{Minimal fraction of appearance of peaks when aligning peaks
with xcms}

\item{minfrac.vote}{Minimal fraction of appearance of fragments when 
generating consensus spectra}

\item{nSlaves}{Number of threads when detecting peaks with xcms}

\item{ppm.pd}{Tolerance in ppm when detecting peaks with xcms}

\item{cutoff}{Peak-Peak Correlation(PPC) threshold to determe if a fragment
is belonging to a precursor ion}

\item{ppm.ms2.mtp}{Tolerance in ppm of the fragment m/z when extracting MS2
EICs with the fragments in multicomplexed spectra}

\item{int.filter.ms1.field}{Field of MS1 peaks after peak detection to decide
if the peak shold be kept}

\item{int.filter.ms1}{Intensity threshold when filtering MS1 peaks}

\item{int.filter.ms2}{Intensity threshold when filtering the fragments in
the multiplexed spectra}

\item{isFWHM}{Logical, if using the full width half maximum of MS1 and MS2
peaks when calculating PPC}

\item{is.plot.eic.feature}{Logical, if plot the EICs of MS1 peaks}

\item{is.plot.eic.spec}{Logical, if plot MS2 EICs}

\item{is.ms1.only}{Logical, if only detect peaks without generating MRM
transitions}

\item{rerun}{Logical, if re-run all data processing works}
}
\description{
An R package to generate MRM transitions from SWATH-MS data.
}
\author{
Yandong Yin \email{yinyandong@sioc.ac.cn}
}
